pyproject.toml
[project]
name = "uvwo"
version = "0.1.0"
description = "Python project using uv"
requires-python = ">=3.11"
dependencies = [
    "langchain-core>=0.3.25",
    "langchain-openai>=0.2.12",
    "langgraph>=0.2.59",
    "pydantic>=2.10.3",
    "python-dotenv>=1.0.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["uvwo"]

[tool.rye]
managed = true
dev-dependencies = []
